refactor: move selectSigner to separate file
